requirements.txt
connexion[swagger-ui]
serverless_wsgi
psycopg2-binary
sqlalchemy
alembic
flask-cors
